+++ common/models/config.go
@@ -113,6 +113,7 @@ type RepoGroupConfig struct {
 	DefaultBranch  string           `toml:"default_branch" json:"default_branch"`
 	BranchSync     string           `toml:"branch_sync" json:"branch_sync"`
 	SyncTags       bool             `toml:"sync_tags" json:"sync_tags"`
+	SyncPRState    bool             `toml:"sync_pr_state" json:"sync_pr_state"`
 	HookPath       string           `toml:"hookpath" json:"hookpath"`
 	CIProvider     string           `toml:"ci_provider" json:"ci_provider"`
 	MergeQueue     MergeQueueConfig `toml:"merge_queue" json:"merge_queue"`