+++ common/models/models.go
@@ -58,6 +58,7 @@ type RepoGroup struct {
 	DefaultBranch  string           `json:"default_branch"`
 	BranchSync     string           `json:"branch_sync"`
 	SyncTags       bool             `json:"sync_tags"`
+	SyncPRState    bool             `json:"sync_pr_state"`
 	HookPath       string           `json:"hookpath"`
 	CIProvider     string           `json:"ci_provider"`
 	MergeQueue     MergeQueueConfig `json:"merge_queue"`